订单微调

+++ modules/payment.go
@@ -1,6 +1,6 @@
 package modules
 
-const(
+const (
 	// PayTypeAliPay 支付类型-支付宝
 	PayTypeAliPay = 1
 	// PayTypeWeChatPay 支付类型-微信
@@ -8,11 +8,11 @@ const(
 )
 
 // Payment 收款
-func Payment(orderID, userID uint64, payType uint, payAccount string, price float32, timestamp string) error {
+func Payment(orderID, userID uint64, payType uint8, payAccount string, price float32, timestamp string) error {
 	return nil
 }
 
 // Refund 退款
-func Refund(orderID, userID uint64, payType uint, payAccount string, price float32, timestamp string) error {
+func Refund(orderID, userID uint64, payType uint8, payAccount string, price float32, timestamp string) error {
 	return nil
 }
